Rename and refactor test_doctest sources
